+++ src/Game/libs/Snake/Systems/MenuKeyboardHandler.hpp
@@ -16,7 +16,7 @@ namespace Snake
         MenuKeyboardHandler();
         ~MenuKeyboardHandler() final = default;
 
-        void run(float deltaTime, Arcade::ECS::IEventManager &eventManager,
+        void run(double deltaTime, Arcade::ECS::IEventManager &eventManager,
             Arcade::ECS::IEntityManager &currentEntityManager) final;
 
         private:
